+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-# Master (will become release 2.8)
+# Master (will become release 2.9)
+
+# Release 2.8
 
 - Python bindings have been moved from the `dune-python` module which is now
   obsolete. To activate Python bindings the CMake flag